Chore #16: seprate model
