+++ feature/home/src/main/java/com/hankki/feature/home/model/StoreItemEntity.kt
@@ -1,0 +1,9 @@
+package com.hankki.feature.home.model
+
+data class StoreItemEntity(
+    val storeImageUrl: String,
+    val category: String,
+    val storeName: String,
+    val price: Int,
+    val heartCount: Int,
+)
